+++ src/HLT_scripts_for_Sergeys_thesis/make_HLT_plots_Sergey.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 from ROOT import TGraphAsymmErrors, TF1, TLegend, TLatex
 from array import array
 from config import CMS
-from tools.ROOT_utililities import set_root_defaults
+from tools.ROOT_utils import set_root_defaults
 import matplotlib.cm as cm
 from matplotlib.ticker import FormatStrFormatter
