+++ src/HLT_scripts_for_Sergeys_thesis/make_jet_response_plots_correction_levels.py
@@ -14,7 +14,7 @@ import matplotlib.pyplot as plt
 from ROOT import TGraphAsymmErrors, TF1
 from array import array
 from config import CMS
-from tools.ROOT_utililities import set_root_defaults
+from tools.ROOT_utils import set_root_defaults
 import matplotlib.cm as cm
 from matplotlib.ticker import FormatStrFormatter
