+++ src/check_CRAB_jobs.py
@@ -1,7 +1,7 @@
 import os as os
 import sys as sys
 import re as re
-from tools.ROOT_utililities import set_root_defaults
+from tools.ROOT_utils import set_root_defaults
 from optparse import OptionParser
 import glob as glob
 from rootpy.io import File
