+++ src/cross_section_measurement/make_fit_variable_plots.py
@@ -7,7 +7,7 @@ from collections import OrderedDict
 from copy import copy, deepcopy
 
 from optparse import OptionParser
-from tools.ROOT_utililities import get_histograms_from_files
+from tools.ROOT_utils import get_histograms_from_files
 from tools.hist_utilities import prepare_histograms, clean_control_region
 from tools.file_utilities import make_folder_if_not_exists
 from tools.plotting import make_data_mc_comparison_plot, Histogram_properties, make_shape_comparison_plot,\
