+++ src/generate_QCD_template_from_data.py
@@ -17,7 +17,7 @@ with the content
 from __future__ import division
 from rootpy.io import File
 from tools.QCD_rate_estimation import estimate_with_fit_to_relative_isolation
-from tools.ROOT_utililities import set_root_defaults
+from tools.ROOT_utils import set_root_defaults
 # one template function per variable
 def get_electron_absolute_eta_templates(b_tag):
     global electron_data_file, met_bins
