+++ src/lepton_scale_factors/measure_2011_hadron_leg.py
@@ -1,7 +1,7 @@
 from optparse import OptionParser
 from read_BLT_ntuple import Particle, read_lepton_collections, match_four_momenta, get_parameters, set_parameter_limits, get_fitted_function_str
 from ROOT import TGraphAsymmErrors, TF1
-from tools.ROOT_utililities import set_root_defaults
+from tools.ROOT_utils import set_root_defaults
 from tools.file_utilities import make_folder_if_not_exists
 import rootpy.plotting.root2matplotlib as rplt
 import matplotlib.pyplot as plt
