+++ tools/HistSet.py
@@ -6,7 +6,7 @@ Created on 8 Jan 2015
 import sys
 from tools.plotting import Histogram_properties, make_shape_comparison_plot,\
     make_data_mc_comparison_plot
-from ROOT_utililities import get_histogram_from_file
+from ROOT_utils import get_histogram_from_file
 
 class HistSet():
     '''
